+++ app/src/main/java/com/techcourse/ManualHandlerAdapter.java
@@ -5,7 +5,7 @@ import jakarta.servlet.http.HttpServletResponse;
 
 import com.interface21.webmvc.servlet.ModelAndView;
 import com.interface21.webmvc.servlet.mvc.asis.Controller;
-import com.interface21.webmvc.servlet.mvc.tobe.HandlerAdapter;
+import com.interface21.webmvc.servlet.mvc.tobe.adapter.HandlerAdapter;
 import com.interface21.webmvc.servlet.view.JspView;
 
 public class ManualHandlerAdapter implements HandlerAdapter {
